@@ -1,17 +1,17 @@
 package json
 
 // https://github.com/ksonnet/ksonnet/blob/master/pkg/kubecfg/diff.go
-func removeFields(config, live interface{}) interface{} {
+func removeFields(config, live any) any {
 	switch c := config.(type) {
-	case map[string]interface{}:
-		l, ok := live.(map[string]interface{})
+	case map[string]any:
+		l, ok := live.(map[string]any)
 		if ok {
 			return RemoveMapFields(c, l)
 		} else {
 			return live
 		}
-	case []interface{}:
-		l, ok := live.([]interface{})
+	case []any:
+		l, ok := live.([]any)
 		if ok {
 			return RemoveListFields(c, l)
 		} else {
@@ -20,12 +20,11 @@ func removeFields(config, live interface{}) interface{} {
 	default:
 		return live
 	}
-
 }
 
 // RemoveMapFields remove all non-existent fields in the live that don't exist in the config
-func RemoveMapFields(config, live map[string]interface{}) map[string]interface{} {
-	result := map[string]interface{}{}
+func RemoveMapFields(config, live map[string]any) map[string]any {
+	result := map[string]any{}
 	for k, v1 := range config {
 		v2, ok := live[k]
 		if !ok {
@@ -39,10 +38,10 @@ func RemoveMapFields(config, live map[string]interface{}) map[string]interface{}
 	return result
 }
 
-func RemoveListFields(config, live []interface{}) []interface{} {
+func RemoveListFields(config, live []any) []any {
 	// If live is longer than config, then the extra elements at the end of the
 	// list will be returned as-is so they appear in the diff.
-	result := make([]interface{}, 0, len(live))
+	result := make([]any, 0, len(live))
 	for i, v2 := range live {
 		if len(config) > i {
 			if v2 != nil {
