@@ -4,6 +4,7 @@ package kube
 import (
 	"bytes"
 	"context"
+	"errors"
 	"fmt"
 	"io"
 	"regexp"
@@ -12,9 +13,9 @@ import (
 
 	"github.com/go-logr/logr"
 
-	v1 "k8s.io/api/core/v1"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
-	apierr "k8s.io/apimachinery/pkg/api/errors"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -88,8 +89,8 @@ func GetResourceKey(obj *unstructured.Unstructured) ResourceKey {
 	return NewResourceKey(gvk.Group, gvk.Kind, obj.GetNamespace(), obj.GetName())
 }
 
-func GetObjectRef(obj *unstructured.Unstructured) v1.ObjectReference {
-	return v1.ObjectReference{
+func GetObjectRef(obj *unstructured.Unstructured) corev1.ObjectReference {
+	return corev1.ObjectReference{
 		UID:        obj.GetUID(),
 		APIVersion: obj.GetAPIVersion(),
 		Kind:       obj.GetKind(),
@@ -102,17 +103,17 @@ func GetObjectRef(obj *unstructured.Unstructured) v1.ObjectReference {
 func TestConfig(config *rest.Config) error {
 	kubeclientset, err := kubernetes.NewForConfig(config)
 	if err != nil {
-		return fmt.Errorf("REST config invalid: %s", err)
+		return fmt.Errorf("REST config invalid: %w", err)
 	}
 	_, err = kubeclientset.ServerVersion()
 	if err != nil {
-		return fmt.Errorf("REST config invalid: %s", err)
+		return fmt.Errorf("REST config invalid: %w", err)
 	}
 	return nil
 }
 
 // ToUnstructured converts a concrete K8s API type to a un unstructured object
-func ToUnstructured(obj interface{}) (*unstructured.Unstructured, error) {
+func ToUnstructured(obj any) (*unstructured.Unstructured, error) {
 	uObj, err := runtime.NewTestUnstructuredConverter(equality.Semantic).ToUnstructured(obj)
 	if err != nil {
 		return nil, err
@@ -121,7 +122,7 @@ func ToUnstructured(obj interface{}) (*unstructured.Unstructured, error) {
 }
 
 // MustToUnstructured converts a concrete K8s API type to a un unstructured object and panics if not successful
-func MustToUnstructured(obj interface{}) *unstructured.Unstructured {
+func MustToUnstructured(obj any) *unstructured.Unstructured {
 	uObj, err := ToUnstructured(obj)
 	if err != nil {
 		panic(err)
@@ -180,7 +181,7 @@ func IsCRD(obj *unstructured.Unstructured) bool {
 // See: https://github.com/ksonnet/ksonnet/blob/master/utils/client.go
 func ServerResourceForGroupVersionKind(disco discovery.DiscoveryInterface, gvk schema.GroupVersionKind, verb string) (*metav1.APIResource, error) {
 	// default is to return a not found for the requested resource
-	retErr := apierr.NewNotFound(schema.GroupResource{Group: gvk.Group, Resource: gvk.Kind}, "")
+	retErr := apierrors.NewNotFound(schema.GroupResource{Group: gvk.Group, Resource: gvk.Kind}, "")
 	resources, err := disco.ServerResourcesForGroupVersion(gvk.GroupVersion().String())
 	if err != nil {
 		return nil, err
@@ -189,11 +190,10 @@ func ServerResourceForGroupVersionKind(disco discovery.DiscoveryInterface, gvk s
 		if r.Kind == gvk.Kind {
 			if isSupportedVerb(&r, verb) {
 				return &r, nil
-			} else {
-				// We have a match, but the API does not support the action
-				// that was requested. Memorize this.
-				retErr = apierr.NewMethodNotSupported(schema.GroupResource{Group: gvk.Group, Resource: gvk.Kind}, verb)
 			}
+			// We have a match, but the API does not support the action
+			// that was requested. Memorize this.
+			retErr = apierrors.NewMethodNotSupported(schema.GroupResource{Group: gvk.Group, Resource: gvk.Kind}, verb)
 		}
 	}
 	return nil, retErr
@@ -215,7 +215,7 @@ func cleanKubectlOutput(s string) string {
 	s = kubectlErrOutRegexp.ReplaceAllString(s, "")
 	s = kubectlErrOutMapRegexp.ReplaceAllString(s, "")
 	s = kubectlApplyPatchErrOutRegexp.ReplaceAllString(s, "")
-	s = strings.Replace(s, "; if you choose to ignore these errors, turn validation off with --validate=false", "", -1)
+	s = strings.ReplaceAll(s, "; if you choose to ignore these errors, turn validation off with --validate=false", "")
 	return s
 }
 
@@ -315,7 +315,7 @@ func SplitYAML(yamlData []byte) ([]*unstructured.Unstructured, error) {
 	for _, yml := range ymls {
 		u := &unstructured.Unstructured{}
 		if err := yaml.Unmarshal([]byte(yml), u); err != nil {
-			return objs, fmt.Errorf("failed to unmarshal manifest: %v", err)
+			return objs, fmt.Errorf("failed to unmarshal manifest: %w", err)
 		}
 		objs = append(objs, u)
 	}
@@ -334,10 +334,10 @@ func SplitYAMLToString(yamlData []byte) ([]string, error) {
 	for {
 		ext := runtime.RawExtension{}
 		if err := d.Decode(&ext); err != nil {
-			if err == io.EOF {
+			if errors.Is(err, io.EOF) {
 				break
 			}
-			return objs, fmt.Errorf("failed to unmarshal manifest: %v", err)
+			return objs, fmt.Errorf("failed to unmarshal manifest: %w", err)
 		}
 		ext.Raw = bytes.TrimSpace(ext.Raw)
 		if len(ext.Raw) == 0 || bytes.Equal(ext.Raw, []byte("null")) {
@@ -407,13 +407,56 @@ func GetDeploymentReplicas(u *unstructured.Unstructured) *int64 {
 	return &val
 }
 
+func GetResourceImages(u *unstructured.Unstructured) []string {
+	var containers []any
+	var found bool
+	var err error
+	var images []string
+
+	containerPaths := [][]string{
+		// Resources without template, like pods
+		{"spec", "containers"},
+		// Resources with template, like deployments
+		{"spec", "template", "spec", "containers"},
+		// Cronjobs
+		{"spec", "jobTemplate", "spec", "template", "spec", "containers"},
+	}
+
+	for _, path := range containerPaths {
+		containers, found, err = unstructured.NestedSlice(u.Object, path...)
+		if found && err == nil {
+			break
+		}
+	}
+
+	if !found || err != nil {
+		return nil
+	}
+
+	for _, container := range containers {
+		containerMap, ok := container.(map[string]any)
+		if !ok {
+			continue
+		}
+
+		image, found, err := unstructured.NestedString(containerMap, "image")
+		if !found || err != nil {
+			continue
+		}
+
+		images = append(images, image)
+	}
+
+	return images
+}
+
 // RetryUntilSucceed keep retrying given action with specified interval until action succeed or specified context is done.
 func RetryUntilSucceed(ctx context.Context, interval time.Duration, desc string, log logr.Logger, action func() error) {
-	pollErr := wait.PollUntilContextCancel(ctx, interval, true, func(ctx context.Context) (bool /*done*/, error) {
-		log.V(1).Info(fmt.Sprintf("Start %s", desc))
+	pollErr := wait.PollUntilContextCancel(ctx, interval, true, func(_ context.Context) (bool /*done*/, error) {
+		log.V(1).Info("Start " + desc)
 		err := action()
 		if err == nil {
-			log.V(1).Info(fmt.Sprintf("Completed %s", desc))
+			log.V(1).Info("Completed " + desc)
 			return true, nil
 		}
 		log.V(1).Info(fmt.Sprintf("Failed to %s: %+v, retrying in %v", desc, err, interval))
@@ -421,6 +464,6 @@ func RetryUntilSucceed(ctx context.Context, interval time.Duration, desc string,
 	})
 	if pollErr != nil {
 		// The only error that can happen here is wait.ErrWaitTimeout if ctx is done.
-		log.V(1).Info(fmt.Sprintf("Stop retrying %s", desc))
+		log.V(1).Info("Stop retrying " + desc)
 	}
 }
