@@ -2,9 +2,10 @@ package kube
 
 import (
 	"fmt"
+	"sort"
+
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/kube-openapi/pkg/util/proto"
-	"sort"
 )
 
 /**
@@ -154,7 +155,7 @@ func parseGroupVersionKind(s proto.Schema) []schema.GroupVersionKind {
 	}
 
 	// gvk extension must be a list of at least 1 element.
-	gvkList, ok := gvkExtension.([]interface{})
+	gvkList, ok := gvkExtension.([]any)
 	if !ok {
 		return []schema.GroupVersionKind{}
 	}
@@ -162,7 +163,7 @@ func parseGroupVersionKind(s proto.Schema) []schema.GroupVersionKind {
 	for _, gvk := range gvkList {
 		// gvk extension list must be a map with group, version, and
 		// kind fields
-		gvkMap, ok := gvk.(map[interface{}]interface{})
+		gvkMap, ok := gvk.(map[any]any)
 		if !ok {
 			continue
 		}
