@@ -5,17 +5,15 @@ var (
 	_ Span   = nopSpan{}
 )
 
-type NopTracer struct {
-}
+type NopTracer struct{}
 
-func (n NopTracer) StartSpan(operationName string) Span {
+func (n NopTracer) StartSpan(_ string) Span {
 	return nopSpan{}
 }
 
-type nopSpan struct {
-}
+type nopSpan struct{}
 
-func (n nopSpan) SetBaggageItem(key string, value interface{}) {
+func (n nopSpan) SetBaggageItem(_ string, _ any) {
 }
 
 func (n nopSpan) Finish() {
